Build fixes for windows * -Wno-braced-scalar-init isn't a valid build flag in MSVC * missing lib dependancy on opencv flann lib * Fixed assorted warnings with new HMD code (mis-typed enum comparisons, etc) * Added virtual destructor to IDeviceInterface to fix long standing warnings

+++ src/psmoveservice/Device/View/ServerHMDView.cpp
@@ -567,7 +567,7 @@ const struct CommonHMDState * ServerHMDView::getState(
 {
     const struct CommonDeviceState *device_state = m_device->getState(lookBack);
     assert(device_state == nullptr ||
-        (device_state->DeviceType >= CommonDeviceState::HeadMountedDisplay &&
+        ((int)device_state->DeviceType >= (int)CommonDeviceState::HeadMountedDisplay &&
         device_state->DeviceType < CommonDeviceState::SUPPORTED_HMD_TYPE_COUNT));
 
     return static_cast<const CommonHMDState *>(device_state);
